+++ firebase-crashlytics/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [changed] Updated `firebase-sessions` dependency to v3.0.4
+
 # 20.0.3
 
 - [fixed] Made creating DataStore files more resilient [#7440]